Radula 0.6 release notes
- cat
Added radula command `cat`, which will print the contents of a remote key over stdout.

- long-keys
Added `keys` flag `-l,--long-keys`, which will prepend the bucket name to the output. This makes it *radula-friendly* when looping on
keys and performing additional operations

- rm multiple keys
radula command `rm` now accepts any number of keys and supports globbing (when quoted to avoid shell expansion)

- info multiple keys/buckets
radula command `info` now accepts any number of keys or bucket names and supports globbing (when quoted to avoid shell expansion).
Additionally, the format of the output has been made to always return a json list, and include an identifier with the info parcel, so
that you know for what subject the details belong.

- resumable uploads / copies
Multipart uploads and streaming copies can now be resumed if interrupted or otherwise not entirely successful. Local uploads will test
the checksums of each existing part against the local file before determining if the part upload can be skipped. Remote-to-remote checking
is more lax, as the source object cannot be parted where it sits.

- configurable chunk sizes
Added radula flag `-c` to set a desired chunk size when uploading or copying. These use a string format with valid values like `15M` and `100mib`.
util-test.py has plently of examples ofthis. The part sizes are stored in the remote object's metadata for use in later checksum verifications.
Verifications on older objects (uploaded with radula < 0.6) will use a legacy strategy. The intent is to be backwards compatible.

- default chunking strategy
The multipart chunking strategy has been simplified to use 100MB part sizes by default.
As mentioned in the previous item, part sizes can be decided when an object is uploaded.
Any part size value will be stored in metadata and be used during later verification.

- object metadata
Uploaded objects will now record the following information to the objects metadata IF the file was multiparted.

    radula-version : the pip version of radula
    radula-chunk-size : part size in bytes
    radula-num-parts : the number of parts transported

- catching S3Response errors
Exceptions from `boto` are more consistent, and print the name of the problem item.

- verification progress
Local checksums now have a progress announcement; helpful for extremely large files.
Overall, the progress indicators have a common interface.

- logging
log message now include a timestamp